fix: remove nor-update and rename fix for corrupted contract
This changes name for fix package that recovers corrupted contract
variable.
It primarilly removes nor-update from it. It can be replaced by just
installing nor-update package.
It is problematic to detect Turris Shield before this fix so it is
easier to install this packages on all moxes. It is also beneficial to
do so because of Omnias that are under contract. We do not want to force
nor-update on all routers so we can't do nor-update here.