foris-controller-openvpn_client: Update and cleanup
Update to the latest version containing few important fixes, but mainly,
NATing and firewall management is no longer patched in, but handled by
vpn_client module and firewall is setup by uci-default script that is no
longer patched in, but included in the package. In the end, this is
mostly about cleaning up the package and adding latest bugfixes.